// src/api/openapi/models/job.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobStatus } from './jobStatus';
import { UserEssential } from './userEssential';
import { Address } from './address';
import { OrderableType } from './orderableType';
import { Client } from './client';
import { Lock } from './lock';


export interface Job { 
    address: Address;
    archive: boolean;
    client: Client;
    code: string;
    completion: string;
    description: string;
    displayable_name: string;
    id: number;
    is_main: boolean;
    is_mini: boolean;
    is_sub: boolean;
    lock: Lock;
    main_job_id: number | null;
    name: string;
    path: string;
    responsible: UserEssential;
    status: JobStatus;
    sub_jobs: Array<Job>;
    timestamp: string;
    type: OrderableType;
    year: number;
}
export namespace Job {
}