// src/api/openapi/models/jobUpdate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressCreate } from './addressCreate';


export interface JobUpdate { 
    address: AddressCreate;
    completion: string;
    description: string;
    name: string;
    responsible_id: number;
}